// apps/backend/src/facturacion/enums/estado-comprobante.enum.ts
export enum EstadoComprobante {
  borrador = 'borrador',
  emitido = 'emitido',
  pendiente_arca = 'pendiente_arca',
  pendiente_posnet = 'pendiente_posnet',
  pendiente_qr = 'pendiente_qr',
  pendiente_transferencia_mp = 'pendiente_transferencia_mp',
  error_arca = 'error_arca',
  anulado = 'anulado',
  importado = 'importado',
}